Preparing the set-up to use Maven
The new file BBCodeManager.js implements the BBCode management and
replacement code. It doesn't do much by itself, one has to develop
extensions that declare themselves to the instance. I'll document that
later.

The other files have been moved out of version control. They are due to
be replaced with cleaner code, which I'll be working on.