Now using Maven, including JSLint and unit test plugins
I added the Maven POM and asked git to ignore the target folder. Also,
following suggestions from JSLint I updated the BBCodeManager class
definition.

Note that I need to think a bit about how the scripts are to be split
and work together. I may end up creating 3 projects: one for an
interface to managing InvisionFree content, one with the current BBCode
manager (that is portable), and one with extensions using that manager
(each extension probably as a module of a bigger project).

I need to see how I can document the Javascript code better, also make
sure it is thoroughly tested (not the case right now).